requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.0
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
django-polymorphic
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1